+++ src/queue/use-queue.ts
@@ -7,7 +7,6 @@ import { useMobile } from './use-mobile';
 import useDataNav, { Opts as UseDataNav } from './use-data-nav';
 import useKeyNav from './use-key-nav';
 import { useListState } from './use-list';
-import useSplit, { SplitOpts } from './use-split';
 import useTabs, { Options as UseTabsOptions } from './use-tabs';
 import useUpdates from './use-updates';
 
@@ -93,7 +92,6 @@ interface Opts<I>
 	 */
 	api?: (id: string, item: I) => string;
 	id?: (i: I) => string;
-	split?: SplitOpts;
 }
 
 const useQueue = <I>({
@@ -103,7 +101,6 @@ const useQueue = <I>({
 	idHashParam,
 	fallback,
 	onActivate,
-	split,
 	...thru
 }: Opts<I>) => {
 	const { items: _items, selected } = thru,
@@ -136,7 +133,6 @@ const useQueue = <I>({
 			[item, id, items],
 		);
 	useKeyNav({ activeTab, ...nav });
-	useSplit({ activeTab, ...split });
 	useUpdates({ ensue, ...nav });
 
 	return {